% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{class}
\name{earthmover_stability-class}
\alias{earthmover_stability-class}
\title{The Earthmover Stability Class}
\description{
This is the type of object returned by the \code{emd_stability()}
function.
}
\section{Slots}{

\describe{
\item{\code{jack_dists}}{the jacknife earthmover distances for samples.}

\item{\code{p}}{the order of the distances.}
}}

