% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{class}
\name{earthmover_stability_summary-class}
\alias{earthmover_stability_summary-class}
\title{the Earthmover Stability Summary Class}
\description{
This is the type of object returned by the
\code{emd_stability_summary()} function.
}
\section{Slots}{

\describe{
\item{\code{ems}}{an object of type "earthmover_stability".}

\item{\code{ks}}{the \code{ks.test} object testing whether \code{x} and \code{y} are different.}

\item{\code{within_p_x}}{the samples from \code{x} that are significanly different
than others with respect to \code{x}.}

\item{\code{within_p_y}}{the samples from \code{y} that are significanly different
than others with respect to \code{y}.}

\item{\code{omnibus_p}}{the samples from \code{x} and \code{y} that are significanly
different than others with respect to \code{x} and \code{y} (the omnibus).}

\item{\code{p_val_thresh}}{the thresholdhold at or below which samples are
deemed significant.}
}}

\seealso{
\code{\link[=emd_stability]{emd_stability()}}
}
