% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{methods}
\name{embedded_emd}
\alias{embedded_emd}
\alias{embedded_emd,data.frame,data.frame,rfsrc,numeric_or_missing-method}
\title{The Earthmover Distance in a Model Embedding}
\usage{
embedded_emd(x, y, model, p)
}
\arguments{
\item{x}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{y}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{model}{the model that will induce the embedded
representation of \code{x} and \code{y}.}

\item{p}{an exponent for the order of the distance (default: 2)}
}
\value{
a list with the following elements:
\itemize{
\item dist - the earthmover distance between \code{x} and \code{y}.
\item pairs - the transportation plan from rows of \code{x} to rows of \code{y}
along with the amount of mass to move.
}
}
\description{
Calculate the earthmover distance between data sets \code{x} and
\code{y} in the embedding defined by \code{model}.
}
\examples{
library(dplyr)
library(randomForestSRC)

# Subset iris for our x and y.
iris1 = iris |> filter(Sepal.Length < 5.8)
iris2 = iris |> filter(Sepal.Length >= 5.8)

# Create a model.
fit1 = rfsrc(
  Species ~ .,
  data = iris1
)

# Calculate the embedding distance between the data sets.
embedded_emd(iris1, iris2, fit1)
}
