% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{methods}
\name{embedded_emd_stability}
\alias{embedded_emd_stability}
\alias{embedded_emd_stability,data.frame,data.frame,rfsrc,numeric_or_missing,logical_or_missing-method}
\title{Stability of the Embedded Earthmove Distance in Each Sample}
\usage{
embedded_emd_stability(x, y, model, p, progress)
}
\arguments{
\item{x}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{y}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{model}{the learner that will induce the embedded}

\item{p}{an exponent for the order of the distance (default: 2).}

\item{progress}{Should the progress be shown as the calculation is being
performed (default: \code{interactive()})?}
}
\value{
a tibble with the following variable.
\itemize{
\item var: The variable (either "x" or "y").
\item row: The row for the corresponding variable.
\item diff: The difference between the earthmover distance and the earthmover
distance with the row removed.
\item p_overall: the empirical p-value of the difference compared to all other
differences.
\item p_within: the empirical p-value of the differences compared to other
difference in the same data set.
\item p_across: the empirical p-value of the difference compared to differences
in the other data set.
}
}
\description{
The earthmover distance is the sum of the Minkowski
distances of samples, weighted their transport coefficients. The
\code{embedded_emd_stability()} function evaluates how stable the distance
is in each of the samples embedding defined by a supvervised learner by
iteratively removing each sample, calating the new earthmover distance and
subtracting that from the earthmover distance
will all samples. The difference between those distances can then be used
to evaluate both how dis-simmilar the sample is to other samples as well
as how much the distance is dependent on the sample.
}
\examples{
library(dplyr)
library(randomForestSRC)

# Subset iris for our x and y.
iris1 = iris |> filter(Sepal.Length < 5.8)
iris2 = iris |> filter(Sepal.Length >= 5.8)

# Create a model.
fit1 = rfsrc(
  Species ~ .,
  data = iris1
)

# Calculate the embedding distance between the data sets.
embedded_emd_stability(iris1, iris2, fit1)
}
