% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{methods}
\name{emd}
\alias{emd}
\alias{emd,data.frame,data.frame,numeric_or_missing-method}
\alias{emd,matrix,matrix,numeric_or_missing-method}
\alias{emd,Matrix,Matrix,numeric_or_missing-method}
\title{The Earthmover Distance}
\usage{
emd(x, y, p)
}
\arguments{
\item{x}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{y}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{p}{an exponent for the order of the distance (default: 2)}
}
\value{
a list with the following elements:
\itemize{
\item dist - the earthmover distance between \code{x} and \code{y}.
\item pairs - the transportation plan from rows of \code{x} to rows of \code{y}
along with the amount of mass to move.
}
}
\description{
Calculate the earthmover distance between data sets \code{x} and
\code{y}.
}
\examples{
X = matrix(rnorm(3*2, mean=-1),ncol=2) # m obs. for X
Y = matrix(rnorm(5*2, mean=+1),ncol=2) # n obs. for Y
emd(X, Y)
}
