% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\docType{methods}
\name{emd_stability}
\alias{emd_stability}
\alias{emd_stability,data.frame,data.frame,numeric_or_missing,logical_or_missing-method}
\alias{emd_stability,matrix,matrix,numeric_or_missing,logical_or_missing-method}
\alias{emd_stability,Matrix,Matrix,numeric_or_missing,logical_or_missing-method}
\title{Stability of the Earthmove Distance in Each Sample}
\usage{
emd_stability(x, y, p, progress)
}
\arguments{
\item{x}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{y}{a \code{matrix}, \code{Matrix}, or \code{data.frame}.}

\item{p}{an exponent for the order of the distance (default: 2).}

\item{progress}{Should the progress be shown as the calculation is being
performed (default: \code{interactive()})?}
}
\value{
a tibble with the following variable.
\itemize{
\item var: The variable (either "x" or "y").
\item row: The row for the corresponding variable.
\item dist: The earthmover distance from the data set, minus the sample to the
other data set.
\item diff: The difference between the earthmover distance and the earthmover
distance with the row removed.
\item p_overall: the empirical p-value of the difference compared to all other
differences.
\item p_within: the empirical p-value of the differences compared to other
difference in the same data set.
\item p_across: the empirical p-value of the difference compared to differences
in the other data set.
}
}
\description{
The earthmover distance is the sum of the Minkowski
distances of samples, weighted their transport coefficients. The
\code{emd_stability()} function evaluates how stable the distance is in each
of the samples by iteratively removing each sample, calating the new
earthmover distance and subtracting that from the earthmover distance
will all samples. The difference between those distances can then be used
to evaluate both how dis-simmilar the sample is to other samples as well
as how much the distance is dependent on the sample.
}
\examples{
X = matrix(rnorm(3*2, mean=-1),ncol=2) # m obs. for X
Y = matrix(rnorm(5*2, mean=+1),ncol=2) # n obs. for Y
emd_stability(X, Y)
}
